feat: Search feature implemented
Modified the BooksGrid and Book components to be more flexible and to be able to use them in the Search component.
Introduced the search feature using the onChange event of the text input and waiting for the user to stop typing befor sending the actual request.
Added a generic book cover image file to be used when no imagem can be retrieved.