feature: Side Menu and Book Details Page
In this commit I created the side menu of the app, wich is activated by the button in the top left corner, on the AppBar component. This menu contains a Profile area with informations of the user of the app and a small navigation menu that leads to the pages of the app.

Also in this commit is the Book Details page. Now when a user clicks in the cover of the book in any page, he is taken to the Book Details page wich shows more detailed info on the book as well as provide a link for the user to preview it.